Moves out OffsetRange
